pyproject.toml
[tool.poetry]
name = "dmarc-report-notifier"
version = "0.0.0"
description = "Headless periodic DMARC report handler"
license = "GPL-3.0-only"
authors = ["Andrew Kvalheim <andrew@kvalhe.im>"]
readme = "README.md"
packages = [{include = "dmarc_report_notifier"}]

[tool.poetry.scripts]
dmarc-report-notifier = "dmarc_report_notifier:main_sync"

[tool.poetry.dependencies]
python = "^3.10"
parsedmarc = "^8.3.2"
matrix-nio = "^0.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
